Implement new brownout plan, thread based updating
